fix: fix delete & bacspace function, add ES6
