+++ index.html
@@ -9,7 +9,7 @@
 
 </head>
 <body id="app">
-    <script src="script.js"></script>
+    <script type="module" src="script.js"></script>
     
 </body>
 </html>
