+++ keys.js
@@ -1,0 +1,28 @@
+export const keysEn = [
+  ["`", "1", "2", "3", "4", "5", "6", "7", "8", "9", "0", "-", "=", "Backspace", "Delete"],
+  ["Tab", "q", "w", "e", "r", "t", "y", "u", "i", "o", "p", "[", "]"],
+  ["CapsLock", "a", "s", "d", "f", "g", "h", "j", "k", "l", ";", "Enter"],
+  ["Shift", "z", "x", "c", "v", "b", "n", "m", ",", ".", "▲", "/", "?"],
+  ["Ctrl", "Alt", " ", "◄", "▼", "►"]
+];
+export const keysEnCaps = [
+  ["~", "!", "@", "#", "$", "%", "^", "&", "*", "(", ")", "-", "=", "Backspace", "Delete"],
+  ["Tab", "Q", "W", "E", "R", "T", "Y", "U", "I", "O", "P", "[", "]"],
+  ["CapsLock", "A", "S", "D", "F", "G", "H", "J", "K", "L", ";", "Enter"],
+  ["Shift", "Z", "X", "C", "V", "B", "N", "M", ",", ".", "▲", "/", "?"],
+  ["Ctrl", "Alt", " ", "◄", "▼", "►"]
+];
+export const keysRu = [
+  ["`", "1", "2", "3", "4", "5", "6", "7", "8", "9", "0", "-", "=", "Backspace", "Delete"],
+  ["Tab", "ё", "й", "ц", "у", "к", "е", "н", "г", "ш", "щ", "з", "х", "ъ", "[", "]"],
+  ["CapsLock", "ф", "ы", "в", "а", "п", "р", "о", "л", "д", "ж", "э", ";", "Enter"],
+  ["Shift", "я", "ч", "с", "м", "и", "т", "ь", "б", "ю", ",", "▲", ".", "/", "?"],
+  ["Ctrl", "Alt", " ", "◄", "▼", "►"]
+];
+export const keysRuCaps = [
+  ["~", "!", "@", "#", "$", "%", "^", "&", "*", "(", ")", "-", "=", "Backspace", "Delete"],
+  ["Tab", "Ё", "Й", "Ц", "У", "К", "Е", "Н", "Г", "Ш", "Щ", "З", "Х", "Ъ", "[", "]"],
+  ["CapsLock", "Ф", "Ы", "В", "А", "П", "Р", "О", "Л", "Д", "Ж", "Э", ";", "Enter"],
+  ["Shift", "Я", "Ч", "С", "М", "И", "Т", "Ь", "Б", "Ю", ",", "▲", ".", "/", "?"],
+  ["Ctrl", "Alt", " ", "◄", "▼", "►"]
+];
